requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.4.0
pandas>=1.3.0
seaborn>=0.11.0
streamlit>=1.25.0
Pillow>=8.0.0
scikit-learn>=0.24.0
